pyproject.toml
[project]
name = "langgraph-playground"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "langchain-core>=1.2.5",
    "langchain-openai>=1.1.6",
    "langgraph>=1.0.5",
    "python-dotenv>=1.2.1",
]
